+++ tests/IntegrationTests/Curiosity.Migrations.PostgreSql.IntegrationTests/PostgreSQLProviderIntegrationTests.cs
@@ -36,11 +36,11 @@ namespace Curiosity.Migrations.PostgreSql.IntegrationTests
             Assert.True(isTableExist);
 
             var desiredDbVersion = new DbVersion(1, 0);
-            var state = await provider.GetDbStateSafeAsync(desiredDbVersion);
+            var state = await provider.GetDbStateSafeAsync(desiredDbVersion, false);
             Assert.Equal(DbState.Outdated, state);
 
             await provider.UpdateCurrentDbVersionAsync($"Version {desiredDbVersion.Major}.{desiredDbVersion.Minor}", desiredDbVersion);
-            var currentDbVersion = await provider.GetDbVersionSafeAsync();
+            var currentDbVersion = await provider.GetDbVersionSafeAsync(false);
             
             Assert.NotNull(currentDbVersion);
             Assert.Equal(desiredDbVersion, currentDbVersion.Value);